Refactor unit tests by grouping related tests and relocating regression tests
This commit performs a sweeping cleanup of the Cendol unit test suite:
- Removed multiple standalone `regr_*.rs` files and cleanly moved their contents to their respective feature modules (e.g., `semantic_scope.rs`, `semantic_expr.rs`).
- Grouped related tests together within single `#[test]` functions (like `test_init_range_features`, `test_struct_bitfield_features`) to significantly reduce test runner overhead.
- Ensured all tests correctly resolve and use established test helpers, eliminating repetitive boilerplate.
- Verified that all unit tests still pass perfectly without any regressions in overall code coverage.
