+++ src/tests.rs
@@ -156,18 +156,6 @@ pub mod guardian_vla;
 pub mod guardian_vm_linkage_constraints;
 
 // Regression
-pub mod regr_different_pointer_levels_warning;
-pub mod regr_incompatible_pointer_warning;
-pub mod regr_incomplete_deref_statement;
-pub mod regr_init_range;
-pub mod regr_mixed_sign_comp;
-pub mod regr_nonconstant_array_designator;
-
-pub mod regr_shadowing;
-pub mod regr_struct_bitfield;
-pub mod regr_tag_kind_mismatch;
-pub mod regr_typeof_ternary;
-pub mod regr_union_cast;
 
 // Test Helpers
 pub mod codegen_alias;
@@ -189,10 +177,7 @@ pub mod parser_struct_pragmas;
 pub mod semantic_binary_conversion;
 pub mod semantic_binary_invalid_operands;
 pub mod semantic_cleanup_coverage;
-pub mod semantic_conversions_uncovered;
 pub mod semantic_gnu_local_label;
 pub mod semantic_mir_label_addr;
 pub mod semantic_types_compatible;
-pub mod test_const_eval_coverage;
-pub mod test_const_eval_unary;
 pub mod test_utils;
