+++ src/tests/semantic_types.rs
@@ -801,3 +801,34 @@ fn test_display_array_types() {
     let arr_vla = reg.array_of(int_ty, ArraySize::Variable(node_ref));
     assert_eq!(reg.display_type(arr_vla), "int[*]");
 }
+
+#[test]
+fn test_tag_kind_mismatch() {
+    // --- test_typedef_struct_union_mismatch ---
+    run_fail_with_message(
+        r#"
+        typedef struct _a a;
+        union _a {
+          int x;
+          int y;
+        };
+        "#,
+        "does not match previous declaration",
+    );
+    // --- test_struct_enum_mismatch ---
+    run_fail_with_message(
+        r#"
+        struct _a;
+        enum _a { A };
+        "#,
+        "does not match previous declaration",
+    );
+    // --- test_enum_struct_mismatch ---
+    run_fail_with_message(
+        r#"
+        enum _a { A };
+        struct _a;
+        "#,
+        "does not match previous declaration",
+    );
+}
