+++ src/tests/semantic_usual_arithmetic_conversions.rs
@@ -1,3 +1,6 @@
+use crate::tests::codegen_common::run_c_code_exit_status;
+use crate::tests::test_utils::run_fail_with_message;
+
 use crate::driver::artifact::CompilePhase;
 use crate::tests::test_utils::run_pipeline;
 
@@ -45,4 +48,52 @@ fn test_usual_arithmetic_conversions() {
     "#;
     let (_, result_long_long_unsigned_long) = run_pipeline(source_long_long_unsigned_long, CompilePhase::Mir);
     assert!(result_long_long_unsigned_long.is_ok());
+}
+
+#[test]
+fn test_usual_arithmetic_conversions_non_arithmetic() {
+    let source = r#"
+        int main() {
+            struct A { int x; } a;
+            1 ? a : 5;
+            return 0;
+        }
+    "#;
+    run_fail_with_message(source, "type mismatch: expected struct A, found int");
+}
+
+#[test]
+fn test_mixed_sign_comparison() {
+    // --- test_mixed_sign_comparison_long_long ---
+    // This tests the case where 'unsigned long' (u64) is compared with 'long long' (i64).
+    // The C11 rule for mixed signedness with rank(st) > rank(ut) is:
+    // If st can represent all values of ut, st wins.
+    // Otherwise, both to unsigned version of st.
+    // On x86-64, rank(LL) = 6, rank(UL) = 5. Sizes are both 64 bits.
+    // So 'unsigned long long' should win.
+    let source = r#"
+        #include <stdint.h>
+        int main() {
+            uint64_t u = 0xFFFFFFFFFFFFFFFBULL; // -5 if signed
+            int64_t s = 1LL;
+            if (u < s) return 1; // Fails if signed comparison
+            return 0; // Passes if unsigned comparison
+        }
+    "#;
+    assert_eq!(run_c_code_exit_status(source), 0);
+    // --- test_mixed_sign_comparison_int_long ---
+    // On x86-64, rank(Long) = 5, rank(UInt) = 4.
+    // sizeof(Long) = 8, sizeof(UInt) = 4.
+    // Signed long (st) can represent all values of unsigned int (ut).
+    // So common type should be 'long' (signed!).
+    let source = r#"
+        #include <stdint.h>
+        int main() {
+            uint32_t u = 0xFFFFFFFFU;
+            int64_t s = 1LL;
+            if (u < s) return 1;
+            return 0;
+        }
+    "#;
+    assert_eq!(run_c_code_exit_status(source), 0);
 }
